+++ src/class_loader.rs
@@ -1,6 +1,6 @@
 use crate::binary::read_binary_file;
 use crate::class::Class;
-use crate::java_class::ClassFile;
+use crate::class_file::ClassFile;
 use std::collections::HashMap;
 
 pub struct ClassLoader {}
