+++ src/main.rs
@@ -3,9 +3,9 @@ extern crate core;
 mod binary;
 mod class;
 mod class_attributes;
+mod class_file;
 mod class_loader;
 mod cp_info;
-mod java_class;
 mod jvm;
 mod thread;
